Добавлены эндпоинты на удаление и отмену review request для draft approach и draft protocol

+++ src/main/kotlin/com/jetbrains/life_science/review/request/service/publish/PublishProtocolRequestServiceImpl.kt
@@ -49,6 +49,9 @@ class PublishProtocolRequestServiceImpl(
     @Transactional
     override fun delete(id: Long) {
         val request = get(id)
+        if (request.state != RequestState.CANCELED) {
+            throw RequestImmutableStateException("Can't delete active request")
+        }
         request.reviews.forEach {
             reviewService.deleteReview(it.id)
         }
